+++ openflowjava/openflow-protocol-impl/src/test/java/org/opendaylight/openflowjava/protocol/impl/util/OF10ActionsSerializerTest.java
@@ -47,6 +47,9 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.common.action.rev1
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.common.action.rev150203.actions.grouping.ActionBuilder;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.common.types.rev130731.PortNumber;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.common.types.rev130731.QueueId;
+import org.opendaylight.yangtools.yang.common.Uint16;
+import org.opendaylight.yangtools.yang.common.Uint32;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 /**
  * Unit tests for OF10ActionsSerializer.
@@ -74,8 +77,8 @@ public class OF10ActionsSerializerTest {
         final List<Action> actions = new ArrayList<>();
         OutputActionCaseBuilder caseBuilder = new OutputActionCaseBuilder();
         OutputActionBuilder outputBuilder = new OutputActionBuilder();
-        outputBuilder.setPort(new PortNumber(42L));
-        outputBuilder.setMaxLength(32);
+        outputBuilder.setPort(new PortNumber(Uint32.valueOf(42)));
+        outputBuilder.setMaxLength(Uint16.valueOf(32));
         caseBuilder.setOutputAction(outputBuilder.build());
         ActionBuilder actionBuilder = new ActionBuilder();
         actionBuilder.setActionChoice(caseBuilder.build());
@@ -83,14 +86,14 @@ public class OF10ActionsSerializerTest {
         actionBuilder = new ActionBuilder();
         SetVlanVidCaseBuilder vlanVidCaseBuilder = new SetVlanVidCaseBuilder();
         SetVlanVidActionBuilder vlanVidBuilder = new SetVlanVidActionBuilder();
-        vlanVidBuilder.setVlanVid(15);
+        vlanVidBuilder.setVlanVid(Uint16.valueOf(15));
         vlanVidCaseBuilder.setSetVlanVidAction(vlanVidBuilder.build());
         actionBuilder.setActionChoice(vlanVidCaseBuilder.build());
         actions.add(actionBuilder.build());
         actionBuilder = new ActionBuilder();
         SetVlanPcpCaseBuilder vlanPcpCaseBuilder = new SetVlanPcpCaseBuilder();
         SetVlanPcpActionBuilder vlanPcpBuilder = new SetVlanPcpActionBuilder();
-        vlanPcpBuilder.setVlanPcp((short) 16);
+        vlanPcpBuilder.setVlanPcp(Uint8.valueOf(16));
         vlanPcpCaseBuilder.setSetVlanPcpAction(vlanPcpBuilder.build());
         actionBuilder.setActionChoice(vlanPcpCaseBuilder.build());
         actions.add(actionBuilder.build());
@@ -128,29 +131,29 @@ public class OF10ActionsSerializerTest {
         actionBuilder = new ActionBuilder();
         SetNwTosCaseBuilder tosCaseBuilder = new SetNwTosCaseBuilder();
         SetNwTosActionBuilder tosBuilder = new SetNwTosActionBuilder();
-        tosBuilder.setNwTos((short) 204);
+        tosBuilder.setNwTos(Uint8.valueOf(204));
         tosCaseBuilder.setSetNwTosAction(tosBuilder.build());
         actionBuilder.setActionChoice(tosCaseBuilder.build());
         actions.add(actionBuilder.build());
         actionBuilder = new ActionBuilder();
         SetTpSrcCaseBuilder tpSrcCaseBuilder = new SetTpSrcCaseBuilder();
         SetTpSrcActionBuilder tpSrcBuilder = new SetTpSrcActionBuilder();
-        tpSrcBuilder.setPort(new PortNumber(6653L));
+        tpSrcBuilder.setPort(new PortNumber(Uint32.valueOf(6653)));
         tpSrcCaseBuilder.setSetTpSrcAction(tpSrcBuilder.build());
         actionBuilder.setActionChoice(tpSrcCaseBuilder.build());
         actions.add(actionBuilder.build());
         actionBuilder = new ActionBuilder();
         SetTpDstCaseBuilder tpDstCaseBuilder = new SetTpDstCaseBuilder();
         SetTpDstActionBuilder tpDstBuilder = new SetTpDstActionBuilder();
-        tpDstBuilder.setPort(new PortNumber(6633L));
+        tpDstBuilder.setPort(new PortNumber(Uint32.valueOf(6633)));
         tpDstCaseBuilder.setSetTpDstAction(tpDstBuilder.build());
         actionBuilder.setActionChoice(tpDstCaseBuilder.build());
         actions.add(actionBuilder.build());
         actionBuilder = new ActionBuilder();
         EnqueueCaseBuilder enqueueCaseBuilder = new EnqueueCaseBuilder();
         EnqueueActionBuilder enqueueBuilder = new EnqueueActionBuilder();
-        enqueueBuilder.setPort(new PortNumber(6613L));
-        enqueueBuilder.setQueueId(new QueueId(400L));
+        enqueueBuilder.setPort(new PortNumber(Uint32.valueOf(6613)));
+        enqueueBuilder.setQueueId(new QueueId(Uint32.valueOf(400)));
         enqueueCaseBuilder.setEnqueueAction(enqueueBuilder.build());
         actionBuilder.setActionChoice(enqueueCaseBuilder.build());
         actions.add(actionBuilder.build());
